push for some small changes.

+++ src/siat/courant/event/Event.java
@@ -3,10 +3,10 @@ package siat.courant.event;
 public interface Event {
 	
 	public long getTimestamp();
-	public long getId();
+	public int getId();
 	public String getEventType();
 	
-	public long getAttributeByName(String attributeName);
+	public int getAttributeByName(String attributeName);
 	public String toString();
 
 }
